feature: nsq output type (#148)
* breaking: Remove all support for passing host/port options.

You now must pass an output in uri form or as a socket path.
Updated remaining docs & tests.

`app` now defaults to `numbat`, in the spirit of requiring as
little as possible to operate.

* Start implementing an nsq output.

* Make the nsq output sort of work, maybe.

Replaced Emitter.parseURI with a createClient() function on the
object itself. Updated the racey tests to match. Ugh.

* Tests broken out into separate files. Racey.

Why, I am not yet sure, but it's a bug.

* Use new race-y tests.

* Finish up the nsq emitter, with example & some docs.